Add more .info file anonymisation.
Ensure that any server addresses in the info file are anonymised. Also remove set the collector host ID and the filter to blank. Both might contain information that might identify the host.

Add a test to make sure the above happens, and also that address event count info is anonymised.