Add limit to blocks in output file.
Once the limit it reached, it triggers a file rotation, if one is
possible.

This change makes a small change to file rotation semantics. Previously,
if a rotation was due but the name did not change, the rotation was
reconsidered only after another rotation period. Now if a rotation fails
because the file name did not change, the file pattern is checked every
subsequent second for a change.