Handle PCAP produced by dnscap.
dnscap (https://github.com/DNS-OARC/dnscap) produces PCAP with contents that are link type RAW. These contain IPv4 or IPv6 packets. Presumably this is an attempt to save storing Ethernet headers.

So, if a packet is type RAW, check the top nibble of the first byte. If it is 4 or 6, attempt to decode it as IPv4 or IPv6.

Add a check that some dnscap output is now read by compactor.