Modify time handling to enable ticks per second to be varied.
This is more complex than one might like, as it requires block parameters to be available to a block. As earliest_time may change as the block is populated, there is no choice but to keep it as a time, and convert to secs/ticks when writing the block. Also, timestamps and response delays need to be converted to time based on the block parameters in force for that particular block, so again the block parameters given in the file header must be available to the block.

As part of the work, remove the time primitives from the CBOR encoder/decoder and handle timestamps as C-DNS objects. CBOR time data types are a different thing; it's be to keep the en/decoding confined to things that are defined in CBOR.