Since block parameters now list opcodes we accept, add compactor accept/ignore options alongside those for RR types.
Also, given format 1.0 has only an accept list, the Ignore report in info file is redundant going forward. So only display it if there are entries to display from format 0.5 or previous.

In the case of RR types, a matching RR is simply omitted from the stored query/response.

For opcodes, if we hit one that's not accepted, ignore the entire query/response.