Revise TShark test script and add filtering of GeoIP output.
I had difficulty following what same-tshark-output.sh was doing, so revised it to (a) use the same operation conventions as the other tests, and (b) keep all files generated with meaningful names in the temporary directory and not delete any as we go along. That way I could just comment out deleting the temporary directory and checking what happened in a failure becomes easy.

It turned out that the reason that test was failing was because tshark was adding GeoIP data to some source and/or destination IPs, but NOT consistently. They'd appear in the dump of the original PCAP but not in the regenerated PCAP, or sometimes vice-versa.

So add a filter for GeoIP lines, and voila! The test passes.