Allow reading and writing of prefix-only IP addresses.
In other words, if the number of bytes in a read address is less than a full address, add in zeros. Or, if writing, convert beyond-prefix bits to 0 and remove trailing 0 bytes.

This turned up a problem with the spec in the draft. With addresses relating to address event types, there are no transport flags, so you can't know if the address is supposed to be IPv4 or IPv6. So add a transport flags field to address event counts.

In turn, the next problem is that addresses at the block level become variable length byte strings; you can't turn them into IPAddress objects because at the point you might do that you don't know if they are IPv4 or IPv6 addresses. So change the block layer to deal with them as byte strings, and remove the IPAddressItem type.

The next knock-on effect is with inspector pseudo-anonymisation. This was done by pseudo-anonymising the entire address block, but again this can't now be done. So switch to anonymising when the address is converted to an IPAddress.

Finally, remove the check that all prefixes are full length when block parameters are read.

File writing has code for outputting addresses with less than full length prefixes, but this is not yet linked to external configuration.