Filter GeoIP from tshark output and simplify test a little.
tshark will include GeoIP info in its output, but it's not consistent as to whether it's there, and tests fail because on one run it will be there and on another it won't. So filter it out of the test comparisons.

Also remove ability to specify input files and temp dir. None of the other tests have them.