Make block items optional where appropriate and make defaults available.
Where possible we use the default at the point of collection, but this isn't always possible, notable with table index values where the basic structure holds the index, not the value. We intend to sort these out at the level of the new QR structure for passing to backends.

Unit tests have been adapted to pass, but some of the integration tests fail. We'll fix these up once the new backend interface is in place.