Start revising handling of excluded data by changing handling of Question.
Move all exclusion into the blockcborwriter layer, and revise writing of CBOR to skip fields that don't have a value. To aid in this, add helpers to the encoder that write an index and a value. If presented with an optional value, only write the index and value if there is a value present.