Revise document building.
Splitting common files shared between man pages and user guide broke the substitutions on the shared files. There's an added complication in that Asciidoctor will only include files relative to a single base directory, and if we're building from a base file outside the build directory we won't pick up any build products. So, given that there are other places in the user guide that should be using substitutions, make all user guide inputs into .adoc.in and build the entire guide from the build directory.