Generate an OPT in responses if necessary.
If the original response had an OPT, generate one from defaults if we've not recorded OPT details.
Use empty RDATA, the query EDNS version, the query DO and the query EDNS payload size, since we don't have any of the response details.

Add a test to ensure we do indeed generate the OPT when not present.