Compactor has historically hardwired DNS to port 53. Make it configurable.
The check for port 53 is a simple initial filter to save checking traffic that certainly isn't DNS. Allow a different port to be given in configuration or on the command line, and store it as a private extension field in the collection parameters. Add the port info to the .info files.