Fix problem due to 64bit signed/unsigned issues.
When confronted with a response with a timestamp just before its corresponding query, the new 1.0 code did not correctly output a negative response delay, but a large positive one. This was at root down to an issue with converting the response delay into ticks. The calculation is delta_in_ns * ticks_per_sec / ns_per_sec. These were mixing signed and unsigned 64bit quantities. The C and C++ rule is that if you mix signed and unsigned quantities, the whole expression is evaluated using the next highest data type size. So, for example, if the quantities were int and unsigned, the expression would evaluate as long. Signedness would therefore be preserved. However, if there is no longer datatype, the 'longest' datatype is used. For 64bit quantities, there is no longer datatype, and the longest type is uint64_t. The expression was therefore evaluated as unsigned.

Correct this by ensuring that ticks_per_sec and ns_per_sec are always signed quantities. Also, just in case, add the missing types to the CBOR encoder so that we don't get any surprising promotion issues (hopefully).

Finally fix up the various test gold files. To save having to re-create the 02 and 05 format gold C-DNS files, retain the old gold.pcap as oldfmtgold.pcap for use in those two tests.