Add custom end time field to block preamble.
The end time is the rollover time or, if the file is closed, e.g.
collection interrupted or PCAP conversion finishes, the end time is the
time of the last collected item.

When converting PCAP, we have no idea when collection actually finished,
so that's the best we can do. It's arguable that if we're collecting off
the wire, we should log the current time. I've chosen to remain
consistent with PCAP conversion.

Add reading and writing the custom field, update the master CDDL, and
fix up tests to cope with the new field.