Only give a block end time when it's definitely known.
After discussion, don't try to give a block end time based on the latest
data in the block. Only specify the end time if it's the result of a
known timestamp, which at the current time means that it's because of a
block/file rollover.

Update the test data appropriately.