Fix stats count misrecording.
If a AE event was the first event in a block, the block start stats
count was not recorded at the right place, and the eventual count would
then be incorrect.

So make sure that the block start stats get recorded from the very first
stats received for a block.

Add a test checking the stats are as expected.