Catch I/O exceptions in body of DNSTAP process and rethrow for greater error clarity.
As things stand, interrupting compactor while DNSTAP is in progress
gives a log message about iostream::clear() that's not terribly clear.

Improve clarity by catching such errors and rethrowing as DNSTAP errors.
In the process, use a wrapper class to ensure that stream exceptions are
reset to their original setting regardless of how the processing loop is
exited.