cppcheck - update suppressions.
Including the build directory on the include path for cppcheck so it
finds config.h removes errors. However, it also means it finds the
protobuf generated header, and for some reason cppcheck always hits the
version check #error therein.

The only way round this I've been able to find is to explicitly define a
preprocessor symbol telling us when cppcheck is busy, and skipping the
include in that case.