Revise doc install and dist handling.
As .adoc files produced from .adoc.in are produced by the Makefile and
not configure, they should also be cleaned in a make clean. The dist
tarball should also include the .adoc.in and not the .adoc.

Finally, the dist should include the user guide overview.png, and that
shouldn't be cleaned. But its build directory copy under doc, created
for install, should be cleaned.