Revise signal handle processing to happen on a dedicated thread.
The range of things you can safely do from a signal handler is limited -
see https://en.cppreference.com/w/cpp/utility/program/signal.

As long as we were just interrupting a PCAP sniffer we could probably
get away with it, but interrupting DNSTAP sockets is more questionable.
Also, the handling was getting very hacky.

So replace with a mechanism using a separate thread to monitor for
expected signals, and call handlers from that thread rather from within
a signal handler.

In the process, slightly rework the dnstap interface, passing the sink
only where it's needed and adding a breakloop() method matching that on
the PCAP sniffers.