When reading DNSTAP from the network, signal the reading thread when interrupted.
Because delivery of the interrupting signal is blocked, the read thread
doesn't see it and synchronous reads from the network don't get
interrupted. So compactor just sits there on a SIGINT.

Get around this by sending the read thread a SIGUSR2, which we're not
using elsewhere, handled with an empty handler. This will interrupt
ongoing reads, and compactor will shut down gracefully.