Amend nc usage to work across all targets.
Unfortunately the version of OpenBSD netcat on Xenial is too old to have
-N. It will just hang around after reading from stdin and keep the
socket open. It does have -q, but FreeBSD doesn't.

So try a bodge that hopefully will work everywhere - give nc a second to
do its work, and then kill it. This will probably fail sometimes on
loaded hosts.