Catch malformed packets in DNSTAP and count them.
At some stage we might want to log them to PCAP output, but that'll
require refactoring the DnsTap interface. For now, just count them while
we find out whether this functionality is required.

Unbound seems to send client messages to DNSTAP before checking them for
validity. We do not know if other resolvers do the same.