DNSTAP: Add sink for malformed messages.
Add a new MalformedMessages sink to DNSTAP processing. Implementing that
sink in compactor, counting the messages (and adding them to the count
of packets received, correcting an existing error) and, if writing to
the ignored PCAP is configured, create an Ethernet/UDP packet
representing the message and writing that to the ignored PCAP.

The fake packet ensure that tools e.g. Wireshark will correctly display
the bad message; otherwise you'd lose the source and destination
addresses/ports and have to explicitly tell Wireshare how to decode the
data.